+++ My project/Assets/Scripts/Monster/Monster_Die.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25af3d7e69782874ea13c12b01b48b9b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
